backend/requirements.txt
fastapi
uvicorn
openai
python-docx
PyPDF2
python-multipart
